+++ examples/hello-library/hello.scm
@@ -9,6 +9,7 @@
 ;;; > huski hello.scm
 ;;;
 (import (scheme base)
+        (libs lib2)
         (rename (prefix (libs lib1) test-)))
 
 (test-lib1-hello)