+++ hs-src/Language/Scheme/Core.hs
@@ -69,7 +69,7 @@ import qualified System.Info as SysInfo
 
 -- |husk version number
 version :: String
-version = "3.15.2"
+version = "3.16"
 
 -- |A utility function to display the husk console banner
 showBanner :: IO ()
@@ -982,7 +982,7 @@ apply cont (HFunc aparams avarargs abody aclosure) args =
      else (liftIO $ extendEnv aclosure $ zip (map ((,) varNamespace) aparams) args) >>= bindVarArgs avarargs >>= (evalBody abody)
   where remainingArgs = drop (length aparams) args
         num = toInteger . length
-        evalBody evBody env = evBody env cont (Nil "") Nothing 
+        evalBody evBody env = evBody env cont (Nil "") (Just [])
 {- TODO: may need to handle cases from Func, such as dynamic winders
         case cont of
             Continuation _ (Just (SchemeBody cBody)) (Just cCont) _ cDynWind -> if length cBody == 0