+++ ChangeLog.markdown
@@ -1,3 +1,30 @@
+v3.16
+--------
+
+- Improved import of libraries:
+
+    - Husk now detects cyclic dependencies and throws an error instead of going into an infinite loop.
+    - Each library is only evaluated once during the import process.
+
+- The `Port` data type has been extended to include an optional in-memory buffer:
+ 
+        Port Handle (Maybe Knob)
+
+  These changes are isolated in husk, but if your code uses any `Port` constructors, you would need to change them, EG: `Port _ Nothing`.
+
+- Added an `-i` command line option to `huski`. This option will start the interactive REPL after a file specified on the command line is executed, and has no effect if no file is specified.
+
+- Added the following R<sup>7</sup>RS I/O functions: 
+
+    - `get-output-bytevector`
+    - `get-output-string`
+    - `open-input-bytevector`
+    - `open-input-string`
+    - `open-output-bytevector`
+    - `open-output-string`
+    - `read-string`
+    - `write-string`
+
 v3.15.2
 --------